Update Js.php
CS-5841: Fix for mixed and doubled slashes. Eg: "\/js/app/admin\_shared.js" is now "/js/app/admin/_shared.js"

This was causing Apache on WAMP stack to throw an issue where it couldn't tell if the file was local or remote and would resolve to // as the file url in turn causing the browser to try and resolve a url without protocol.