Update Js.php
CS-5841: Fixed issue where Apache on Windows was unable to render the correct url for generated JS files in the admin due to mixed forward, back and double slashes.

Output: "\/js/app\_shared.js"

The browser was attempting to resolve the malformed URL to "//js/app/_shared.js" which is an incorrect URI without protocol. Have added check to the for each loop which outputs the final addresses to change all "\" to "/" and then to reduce all "//" to "/" as well.

Tested in Windows Apache and Mac Apache which should also be applicable to Linux Apache.